Big refactoring. Now some types and functions are as modular codes
