+++ dynamic/half-dynamic-queue.c
@@ -15,11 +15,20 @@
  * =====================================================================================
 **/
 
-
+// standard headers
 #include <stdio.h>
 #include <stdlib.h>
 #include <time.h>
-#include "clear-pause.h"
+
+// my headers
+#include "clear-pause.h" 
+#include "meta-type.h"
+#include "list-type.h"
+#include "declarations.h"
+#include "meta-functions.h"
+#include "list-functions.h"
+
+// my macros
 #define CLEAR "clear || cls"
 
 
@@ -32,51 +41,6 @@
  * 
  * ============================================
  */
-
-/* ================================================
- *
- *  -*- All structures definition are below  -*-
- *
- *=================================================
- */
-
-typedef union {
-    int integer;
-    char character;
-    float real;
-} meta_thing;
-
-
-typedef enum {
-    integer,
-    character,
-    real,
-} identifier;
-
-
-typedef struct {
-    meta_thing data;
-    identifier type;
-} meta_data;
-
-
-typedef enum {
-    empty,
-    available,
-    full
-} id_state;
-
-
-typedef struct {
-    meta_data *elements;
-    id_state state;
-    int last_index;
-    int size;
-} queue;
-
-// for aux the enums
-static const char *type_strings[] = { "integer", "character",  "real"};
-static const char *state_strings[] = {"empty", "available", "full"};
 
 
 
@@ -94,41 +58,6 @@ static const char *state_strings[] = {"empty", "available", "full"};
 
 
 
-
-/* ==========================================
- *
- *  -*- Declarations of all functions  -*-
- *
- * ==========================================
- */
-
-
-// methods of list
-int step(queue *q);
-int insert(queue *q);
-int search(queue *q, meta_thing thing);
-int erase(queue *q, meta_thing thing);
-int edit(queue *q, int index);
-
-// aux-functions
-void type_choose(queue *q);
-void start(queue *q);
-void random_values(queue *q);
-void rearrange(queue *q);
-void del(queue *q, int index);
-void verify_state(queue *q);
-
-// in-out and terminal functions
-int  generate(const char *message);
-meta_thing new_thing(queue *q);
-void insert_on(queue *q, int index);
-void print_element(meta_data element);
-void print_list(queue *q);
-void menu(queue *q);
-void clear_buffer(void);
-void pause(const char* msg);
-
-
 /* ==========================================
  *
  *  -*-       THE BIG MAIN CODE           -*-
@@ -137,7 +66,7 @@ void pause(const char* msg);
  */
 
 int main(void) {
-    queue q;
+    list q;
 
     start(&q);
     menu(&q);
@@ -148,37 +77,15 @@ int main(void) {
 
 /* ==========================================
  *
- *  -*- Definition of all functions  -*-
+ *  -*- Definition of all main queue functions  -*-
  *
  * ==========================================
  */
 
 
-void type_choose(queue *q) {
-    int command, i;
-    puts("Choose a type for the list");
-    printf("1.Int\n2.Char\n3.Float\n\n");
-    printf("Enter a command: ");
-    scanf("%d", &command);
-    clear_buffer();
-
-    for (i= 0; i < q->size; i++) {
-        //that verification it will usefull on the type change
-        if (q->elements[i].data.integer == -1){
-            if (command == 1)
-               q->elements[i].type = integer;
-            else if (command == 2)
-               q->elements[i].type = character;
-            else if (command == 3)
-               q->elements[i].type = real;
-            else        
-               type_choose(q);
-        }
-    }
-}
 
 // iniciar q
-void start(queue *q) {
+void start(list *q) {
     // get the size of list
     int size = generate("Enter a size: ");
     q->size = size;
@@ -203,48 +110,6 @@ void start(queue *q) {
     type_choose(q);
 }
 
-void random_values(queue *q) {
-    int i;
-    time_t t;
-    srand((unsigned) time(&t));
-
-    for (i = 0; i < q->size; i++) {
-        identifier type = q->elements[i].type;
-        if (type == integer)
-            q->elements[i].data.integer = rand() % 100;
-        else if (type == character)
-            q->elements[i].data.character = 'a' + rand() % 26;
-        else if (type == real)
-            q->elements[i].data.real = (rand() % 100) / (rand () % 100);
-    }
-    
-    q->last_index = i + 1;
-}
-
-void rearrange(queue *q) {
-    // ignoring empty blocks and rearrange the vector
-    int i, j;
-    for (i = 0, j = 0; i < q->size; i++) {
-        if (q->elements[i].data.integer != -1) {
-            q->elements[j] = q->elements[i];
-            j++;
-        }
-    }
-
-    // assigning the garbage in right of vector as empty
-    for (i = j; i < q->size; i++) {
-        q->elements[i].data.integer = -1;
-    }
-
-    q->last_index = j - 1;
-}
-
-// delete some element
-void del(queue *q, int index) {
-    q->elements[index].data.integer = -1;
-    rearrange(q);
-}
-
 
 // generate a int number with a message for stdout
 int generate(const char *message) {
@@ -258,43 +123,8 @@ int generate(const char *message) {
 }
 
 
-meta_thing new_thing(queue *q){
-    printf("Insert a thing: ");
-    identifier type = q->elements[0].type;
-    meta_thing the_thing;
-
-    if (type == integer)
-        scanf("%d", &the_thing.integer);
-    else if (type == character)
-        scanf("%c", &the_thing.character);
-    else if (type == real)
-        scanf("%f", &the_thing.real);
-
-    clear_buffer();
-
-    return the_thing;
-}
-
-
-
-void insert_on(queue *q, int index) {
-    meta_thing the_thing = new_thing(q);
-
-    q->elements[index].data = the_thing;
-}
-
-void verify_state(queue *q) {
-    if (q->last_index == -1)
-        q->state = empty;
-    else if (q->last_index >= q->size - 1)
-        q->state = full;
-    else
-        q->state = available;
-}
-
-
 // step the queue (remove the head)
-int step(queue *q) {
+int step(list *q) {
     if (q->state == empty) {
         return -1;
     } else {
@@ -308,7 +138,7 @@ int step(queue *q) {
 
 
 // insert a value in a tail of queue;
-int insert(queue *q) {
+int insert(list *q) {
     if (q->state == full)
         return -1;
 
@@ -318,7 +148,7 @@ int insert(queue *q) {
     return 0;
 }
 
-int search(queue *q, meta_thing thing) {
+int search(list *q, something thing) {
     int i;
 
     for (i = 0; i < q->size; i++)
@@ -327,7 +157,7 @@ int search(queue *q, meta_thing thing) {
     return -1;
 }
 
-int erase(queue *q, meta_thing thing) {
+int erase(list *q, something thing) {
     int index = search(q, thing);
 
     if (index != -1)
@@ -338,7 +168,7 @@ int erase(queue *q, meta_thing thing) {
     return index;
 }
 
-int edit(queue *q, int index) {
+int edit(list *q, int index) {
     // verify if index is not do the pesar of violate the lenght of vector
     if (!(index >= 0 && index <= q->size))
         return -1;
@@ -348,36 +178,10 @@ int edit(queue *q, int index) {
     return 0;
 }
 
-void print_element(meta_data element) {
-    identifier type = element.type;
 
-    if (element.data.integer != -1) {
-        if (type == integer)
-            printf(" (%d) ", element.data.integer);
-        else if (type == character)
-            printf(" (%c) ", element.data.character);
-        else if (type == real)
-            printf(" (%.2f) ", element.data.real);
-    } else {
-        printf(" () ");
-    }
-}
-
-
-void print_list(queue *q) {
-    int i;
-    identifier type = q->elements[0].type;
-
-    printf("Queue: [ ");
-    for (i = 0; i < q->size; i++)
-        print_element(q->elements[i]);
-    printf("]\n");
-}
-
-
-void menu(queue *q) {
+void menu(list *q) {
     int command, status;
-    meta_thing element;
+    something element;
 
     do {
         system(CLEAR);
@@ -419,7 +223,7 @@ void menu(queue *q) {
 
             case 4:
                 printf("== Search ==\n");
-                element = new_thing(q);
+                element = new_thing(q->elements[0].type);
                 status = search(q, element);
                 
                 // output
@@ -440,7 +244,7 @@ void menu(queue *q) {
 
             case 6:
                 printf("== Erase element ==\n");
-                element = new_thing(q);
+                element = new_thing(q->elements[0].type);
                 status = erase(q, element);
                 
                 // output
